requirements.txt
Django==4.1.5
djangorestframework==3.14.0
psycopg2-binary==2.9.5
python-dotenv==0.21.0
text-unidecode==1.3
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
django-mptt==0.14.0
python-telegram-bot==13.12
django-versatileimagefield==3.0
soapfish2==0.7.1

